# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AutoCare Hub API
  version: 1.0.0
  description: >
    API para gerenciamento de clientes, veículos, serviços, peças, ordens de
    serviço, orçamentos, aprovações e acompanhamento de status de uma oficina
    mecânica.
servers:
  - url: http://localhost:8080
    description: Ambiente local
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/customers/{customerId}/vehicles:
    get:
      tags:
        - Vehicles
      summary: Lista veículos de um cliente
      operationId: listVehiclesByCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerIdPathParam'
      responses:
        '200':
          description: Veículos encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/vehicles:
    post:
      tags:
        - Vehicles
      summary: Cria um veículo
      operationId: createVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'
            examples:
              veiculoClienteSeed:
                summary: Veículo novo vinculado a cliente existente
                value:
                  customerId: 10000000-0000-0000-0000-000000000001
                  plate: TST1A23
                  brand: Honda
                  model: Civic Touring
                  year: 2021
                  mileage: 42000
      responses:
        '201':
          description: Veículo criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags:
        - Vehicles
      summary: Lista veículos
      operationId: listVehicles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Veículos encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/vehicles/{vehicleId}:
    get:
      tags:
        - Vehicles
      summary: Busca um veículo por id
      operationId: getVehicleById
      parameters:
        - $ref: '#/components/parameters/VehicleIdPathParam'
      responses:
        '200':
          description: Veículo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Vehicles
      summary: Atualiza um veículo
      operationId: updateVehicle
      parameters:
        - $ref: '#/components/parameters/VehicleIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleRequest'
            examples:
              atualizarVeiculo:
                summary: Atualiza veículo e mantém ativo
                value:
                  customerId: 10000000-0000-0000-0000-000000000001
                  plate: TST1A23
                  brand: Honda
                  model: Civic Touring
                  year: 2021
                  mileage: 42500
                  active: true
      responses:
        '200':
          description: Veículo atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Vehicles
      summary: Remove logicamente um veículo
      operationId: deleteVehicle
      parameters:
        - $ref: '#/components/parameters/VehicleIdPathParam'
      responses:
        '204':
          description: Veículo removido
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/workshop-services:
    post:
      tags:
        - Workshop Services
      summary: Cria um serviço da oficina
      operationId: createWorkshopService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkshopServiceRequest'
            examples:
              diagnostico:
                summary: Serviço de diagnóstico para catálogo da oficina
                value:
                  name: Diagnostico eletrico avancado
                  description: Verificacao de alternador, bateria, chicote e modulos eletronicos.
                  basePrice: 250.00
                  estimatedTimeInMinutes: 120
      responses:
        '201':
          description: Serviço criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopServiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags:
        - Workshop Services
      summary: Lista serviços da oficina
      operationId: listWorkshopServices
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Serviços encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopServiceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workshop-services/{serviceId}:
    get:
      tags:
        - Workshop Services
      summary: Busca um serviço da oficina por id
      operationId: getWorkshopServiceById
      parameters:
        - $ref: '#/components/parameters/ServiceIdPathParam'
      responses:
        '200':
          description: Serviço encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopServiceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Workshop Services
      summary: Atualiza um serviço da oficina
      operationId: updateWorkshopService
      parameters:
        - $ref: '#/components/parameters/ServiceIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkshopServiceRequest'
            examples:
              atualizarServico:
                summary: Atualiza serviço mantendo ativo
                value:
                  name: Diagnostico eletrico avancado
                  description: Verificacao de alternador, bateria, chicote, fusivel e modulos eletronicos.
                  basePrice: 280.00
                  estimatedTimeInMinutes: 150
                  active: true
      responses:
        '200':
          description: Serviço atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopServiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Workshop Services
      summary: Remove logicamente um serviço da oficina
      operationId: deleteWorkshopService
      parameters:
        - $ref: '#/components/parameters/ServiceIdPathParam'
      responses:
        '204':
          description: Serviço removido
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CustomerIdPathParam:
      name: customerId
      in: path
      required: true
      example: 10000000-0000-0000-0000-000000000001
      schema:
        type: string
        format: uuid
    VehicleIdPathParam:
      name: vehicleId
      in: path
      required: true
      example: 20000000-0000-0000-0000-000000000001
      schema:
        type: string
        format: uuid
    ServiceIdPathParam:
      name: serviceId
      in: path
      required: true
      example: 30000000-0000-0000-0000-000000000004
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Requisição invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 400
            error: Bad Request
            message: Invalid request body
            path: /api/v1/service-orders
            details:
              - customerDocument: must not be blank
    Unauthorized:
      description: Credenciais ausentes ou invalidas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 401
            error: Unauthorized
            message: Missing or invalid credentials
            path: /api/v1/customers
            details: [ ]
...
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 404
            error: Not Found
            message: Resource not found
            path: /api/v1/customers/00000000-0000-0000-0000-000000000000
            details: [ ]
...
  schemas:
...
    CreateVehicleRequest:
      type: object
      required:
        - customerId
        - plate
        - brand
        - model
        - year
        - mileage
      properties:
        customerId:
          type: string
          format: uuid
        plate:
          type: string
          minLength: 7
          maxLength: 8
          pattern: '^[A-Za-z]{3}-?[0-9]{4}$|^[A-Za-z]{3}-?[0-9][A-Za-z][0-9]{2}$'
          example: ABC1D23
        brand:
          type: string
          minLength: 2
          maxLength: 60
        model:
          type: string
          minLength: 1
          maxLength: 80
        year:
          type: integer
          minimum: 1900
          maximum: 2100
        mileage:
          type: integer
          minimum: 0
    UpdateVehicleRequest:
      allOf:
        - $ref: '#/components/schemas/CreateVehicleRequest'
        - type: object
          required:
            - active
          properties:
            active:
              type: boolean
    VehicleResponse:
      type: object
      required:
        - id
        - customerId
        - plate
        - brand
        - model
        - year
        - mileage
        - active
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        plate:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        active:
          type: boolean
    VehicleListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VehicleResponse'
    CreateWorkshopServiceRequest:
      type: object
      required:
        - name
        - description
        - basePrice
        - estimatedTimeInMinutes
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
          minLength: 5
          maxLength: 500
        basePrice:
          type: number
          format: double
          minimum: 0
        estimatedTimeInMinutes:
          type: integer
          minimum: 1
    UpdateWorkshopServiceRequest:
      allOf:
        - $ref: '#/components/schemas/CreateWorkshopServiceRequest'
        - type: object
          required:
            - active
          properties:
            active:
              type: boolean
    WorkshopServiceResponse:
      type: object
      required:
        - id
        - name
        - description
        - basePrice
        - estimatedTimeInMinutes
        - active
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        basePrice:
          type: number
          format: double
        estimatedTimeInMinutes:
          type: integer
        active:
          type: boolean
    WorkshopServiceListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkshopServiceResponse'
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          minimum: 100
          maximum: 599
        error:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        path:
          type: string
          minLength: 1
        details:
          type: array
          items:
            type: string